+++ CMakeLists.txt
@@ -124,9 +124,7 @@ set(libobjc_CXX_SRCS
 # Windows does not use DWARF EH, except when using the GNU ABI (MinGW)
 if (WIN32 AND NOT MINGW)
 	list(APPEND libobjc_CXX_SRCS eh_win32_msvc.cc)
-elseif (MINGW)
-	list(APPEND libobjc_CXX_SRCS eh_win32_mingw.m)
-else ()
+elseif (NOT MINGW)
 	list(APPEND libobjc_C_SRCS eh_personality.c)
 endif ()
 
@@ -230,7 +228,7 @@ if (WIN32 AND NOT MINGW)
 	message(STATUS "Using MSVC-compatible exception model")
 elseif (MINGW)
 	message(STATUS "Using MinGW-compatible exception model")
-	list(APPEND libobjc_CXX_SRCS objcxx_eh.cc)
+	list(APPEND libobjc_CXX_SRCS objcxx_eh.cc objcxx_eh_mingw.cc)
 else ()
 	separate_arguments(EH_PERSONALITY_FLAGS NATIVE_COMMAND ${CMAKE_CXX_FLAGS})
 	if (CMAKE_CXX_COMPILER_TARGET)
